requirements.txt
autopep8==2.0.1
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
pycodestyle==2.10.0
PyJWT==2.6.0
six==1.16.0
SQLAlchemy==2.0.0
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
